Allow to choose the extension for his files #5
The user can use scripts "add_gps_position.bat" and
"retrieve_metadata.bat" with the option "-ext" to specify
the extension to use.

/!\ It's the responsibility of the user to know if the extension
is a valid image format. It's also up to him to know if the
file supports the tag to add (for example gps tag for png are not
working).